Refactor prov search to make it less entwined and more testable
